pytest-tutorial/testcase_parametrize/requirements.txt
allure-pytest==2.8.33
allure-python-commons==2.8.33
apipkg==1.5
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
execnet==1.8.0
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==9.1.1
pytest-xdist==2.2.0
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.0
